# specs/001-boq-price-forecast/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: BoQ Price Forecast API
  version: 1.0.0
  description: |
    REST API for AI-powered BoQ price adjustment and forecasting.
    Supports Excel upload, weight attribution, price recalculation, forecasting, and defense document generation.
  contact:
    name: BoQ Price Forecast Team
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.boqprice.example.com
    description: Production server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JobStatus:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        message:
          type: string
        result:
          type: object
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # ===== BoQ =====
...
    ExportExcelRequest:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        include_weights:
          type: boolean
          default: true
        include_forecasts:
          type: boolean
          default: true
        include_comparison:
          type: boolean
          default: true
      required: [job_id]
    ExportExcelResponse:
      type: object
      properties:
        download_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
        filename:
          type: string
    ExportDefenseDocRequest:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        item_ids:
          type: array
          items:
            type: string
            format: uuid
        chapter_ids:
          type: array
          items:
            type: string
            format: uuid
        format:
          type: string
          enum: [pdf, html]
          default: pdf
      required: [job_id]
    ExportDefenseDocResponse:
      type: object
      properties:
        download_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
        filename:
          type: string

    # ===== Market Data =====
...
paths:
...
  /exports/excel:
    post:
      tags: [Exports]
      summary: Export updated BoQ to Excel
      operationId: exportExcel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportExcelRequest'
      responses:
        '200':
          description: Excel file ready for download
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportExcelResponse'
        '202':
          description: Export job started (large files)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
  /exports/defense-doc:
    post:
      tags: [Exports]
      summary: Generate defense document (سند دفاعیه)
      operationId: exportDefenseDoc
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportDefenseDocRequest'
      responses:
        '200':
          description: Defense document ready for download
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportDefenseDocResponse'
        '202':
          description: Generation job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'

  # ===== Market Data =====
...